perf: convert FlashblockAccessList directly to HashedPostState
