+++ crates/primitives/Cargo.toml
@@ -20,6 +20,8 @@ reth-optimism-primitives.workspace = true
 reth-optimism-node.workspace = true
 reth-primitives-traits.workspace = true
 reth-optimism-chainspec.workspace = true
+reth-trie-common.workspace = true
+reth-storage-api.workspace = true
 alloy-eips.workspace = true
 alloy-eip7928.workspace = true
 alloy-consensus.workspace = true
